+++ crypto/evp/e_xcbc_d.c
@@ -68,7 +68,7 @@
 static int desx_cbc_init_key(EVP_CIPHER_CTX *ctx, const unsigned char *key,
                              const unsigned char *iv, int enc);
 static int desx_cbc_cipher(EVP_CIPHER_CTX *ctx, unsigned char *out,
-                           const unsigned char *in, unsigned int inl);
+                           const unsigned char *in, size_t inl);
 
 typedef struct {
     DES_key_schedule ks;        /* key schedule */
@@ -110,7 +110,7 @@ static int desx_cbc_init_key(EVP_CIPHER_CTX *ctx, const unsigned char *key,
 }
 
 static int desx_cbc_cipher(EVP_CIPHER_CTX *ctx, unsigned char *out,
-                           const unsigned char *in, unsigned int inl)
+                           const unsigned char *in, size_t inl)
 {
     DES_xcbc_encrypt(in, out, inl, &data(ctx)->ks,
                      (DES_cblock *)&(ctx->iv[0]),
